WIP: genomic substitution validator
